add render item with indicators
